+++ ProfessionalNetwork.Domaine/Entities/Notifications.cs
@@ -11,10 +11,10 @@ namespace ProfessionalNetwork.Domaine.Entities
     public class Notifications
     {
         [Key, Column(Order = 0)]
-        public int Id_JobOffer { get; set; }
+        public long Id_JobOffer { get; set; }
 
         [Key, Column(Order = 1)]
-        public int id_jobseeker { get; set; }
+        public long id_jobseeker { get; set; }
         [DataType(DataType.DateTime)]
         public DateTime Date_Notif { get; set; }
         public string Type { get; set; }
