+++ ProfessionalNetwork.Domaine/Entities/Question.cs
@@ -10,7 +10,7 @@ namespace ProfessionalNetwork.Domaine.Entities
     public class Question
     {
 
-        public int Id_Question { get; set; }
+        public long Id_Question { get; set; }
         public string Content_Question { get; set; }
         public int Correct_AnswerID{ get; set; }
         public int Point_Question { get; set; }
